Remove dir when plugin remove
