+++ cogs/utility.py
@@ -24,7 +24,6 @@ from pkg_resources import parse_version
 
 from core import checks
 from core.changelog import Changelog
-from core.decorators import trigger_typing
 from core.models import InvalidConfigError, PermissionLevel
 from core.paginator import EmbedPaginatorSession, MessagePaginatorSession
 from core import utils
@@ -248,7 +247,7 @@ class Utility(commands.Cog):
 
     @commands.command()
     @checks.has_permissions(PermissionLevel.REGULAR)
-    @trigger_typing
+    @utils.trigger_typing
     async def changelog(self, ctx, version: str.lower = ""):
         """Shows the changelog of the Modmail."""
         changelog = await Changelog.from_url(self.bot)
@@ -282,7 +281,7 @@ class Utility(commands.Cog):
 
     @commands.command(aliases=["bot", "info"])
     @checks.has_permissions(PermissionLevel.REGULAR)
-    @trigger_typing
+    @utils.trigger_typing
     async def about(self, ctx):
         """Shows information about this bot."""
         embed = discord.Embed(color=self.bot.main_color, timestamp=datetime.utcnow())
@@ -331,7 +330,7 @@ class Utility(commands.Cog):
 
     @commands.command()
     @checks.has_permissions(PermissionLevel.REGULAR)
-    @trigger_typing
+    @utils.trigger_typing
     async def sponsors(self, ctx):
         """Shows a list of sponsors."""
         resp = await self.bot.session.get(
@@ -352,7 +351,7 @@ class Utility(commands.Cog):
 
     @commands.group(invoke_without_command=True)
     @checks.has_permissions(PermissionLevel.OWNER)
-    @trigger_typing
+    @utils.trigger_typing
     async def debug(self, ctx):
         """Shows the recent application-logs of the bot."""
 
@@ -407,7 +406,7 @@ class Utility(commands.Cog):
 
     @debug.command(name="hastebin", aliases=["haste"])
     @checks.has_permissions(PermissionLevel.OWNER)
-    @trigger_typing
+    @utils.trigger_typing
     async def debug_hastebin(self, ctx):
         """Posts application-logs to Hastebin."""
 
@@ -450,7 +449,7 @@ class Utility(commands.Cog):
 
     @debug.command(name="clear", aliases=["wipe"])
     @checks.has_permissions(PermissionLevel.OWNER)
-    @trigger_typing
+    @utils.trigger_typing
     async def debug_clear(self, ctx):
         """Clears the locally cached logs."""
 
@@ -658,7 +657,7 @@ class Utility(commands.Cog):
 
     @commands.command()
     @checks.has_permissions(PermissionLevel.ADMINISTRATOR)
-    @trigger_typing
+    @utils.trigger_typing
     async def ping(self, ctx):
         """Pong! Returns your websocket latency."""
         embed = discord.Embed(
